requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
cloudscraper==1.2.71
beautifulsoup4==4.12.3
lxml==5.3.0
APScheduler==3.10.4
discord.py==2.4.0
requests==2.32.3
